// frontend/src/api/model/treatmentKind.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Hospital Internment API
 * OpenAPI spec version: 0.1.0
 */

export type TreatmentKind = typeof TreatmentKind[keyof typeof TreatmentKind];


export const TreatmentKind = {
  MEDICATION: 'MEDICATION',
  TREATMENT: 'TREATMENT',
} as const;